Move gateway from api to bl layer
